+++ apps/E2E/src/Drawer/consts.ts
@@ -1,0 +1,8 @@
+export const HOMEPAGE_Drawer_BUTTON = 'Homepage_Drawer_Button';
+export const Drawer_TESTPAGE = 'Drawer_TestPage';
+
+/* E2E Testing Drawer 1 */
+export const Drawer_TEST_COMPONENT = 'Drawer_Test_Component'; // A component on each specific test page
+export const Drawer_TEST_COMPONENT_DISMISS_BUTTON = 'Drawer_Test_Component_Dismiss_Button'; // A component on each specific test page
+
+export const Drawer_TEXT = 'Drawer_Text'; // For testing the text Drawer functionality of Drawer