+++ packages/components/RadioGroup/src/RadioGroup/__tests__/RadioGroupExperimental.test.tsx
@@ -3,8 +3,8 @@ import * as React from 'react';
 import { checkRenderConsistency, checkReRender } from '@fluentui-react-native/test-tools';
 import * as renderer from 'react-test-renderer';
 
-import { RadioGroup } from './RadioGroup';
-import { Radio } from '../Radio/Radio';
+import { Radio } from '../../Radio/Radio';
+import { RadioGroup } from '../RadioGroup';
 
 jest.useFakeTimers();