+++ packages/experimental/Drawer/src/index.ts
@@ -1,0 +1,3 @@
+export { Drawer } from './Drawer';
+export type { DrawerProps, DrawerSlotProps, DrawerTokens, DrawerType, DrawerPositionType } from './Drawer.types';
+export { DrawerName } from './Drawer.types';